fragment activity 구성 및 위젯 위치조정
