+++ swProject/.idea/misc.xml
@@ -22,6 +22,7 @@
         <entry key="..\:/Users/USER/Desktop/OpenSource_SW/swProject/app/src/main/res/layout/soccer.xml" value="0.15670289855072464" />
         <entry key="..\:/Users/USER/Desktop/OpenSource_SW/swProject/app/src/main/res/layout/soccer_news.xml" value="0.15670289855072464" />
         <entry key="..\:/Users/rlagk/swProject/app/src/main/res/drawable/gradient.xml" value="0.1" />
+        <entry key="..\:/Users/rlagk/swProject/app/src/main/res/drawable/ic_baseline_home_24.xml" value="0.1" />
         <entry key="..\:/Users/rlagk/swProject/app/src/main/res/drawable/ic_baseline_sports_baseball_24.xml" value="0.1" />
         <entry key="..\:/Users/rlagk/swProject/app/src/main/res/drawable/ic_baseline_sports_soccer_24.xml" value="0.1" />
         <entry key="..\:/Users/rlagk/swProject/app/src/main/res/drawable/shape_button.xml" value="0.1" />
