+++ swProject/app/src/main/java/com/example/swproject/fragment/fragment_player_soccer.java
@@ -1,19 +1,20 @@
-package com.example.swproject;
+package com.example.swproject.fragment;
 
 import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import androidx.fragment.app.Fragment;
+
+import com.example.swproject.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link fragment_team_soccer#newInstance} factory method to
+ * Use the {@link fragment_player_soccer#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class fragment_team_soccer extends Fragment {
+public class fragment_player_soccer extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +25,7 @@ public class fragment_team_soccer extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public fragment_team_soccer() {
+    public fragment_player_soccer() {
         // Required empty public constructor
     }
 
@@ -34,11 +35,11 @@ public class fragment_team_soccer extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment fragment_team_soccer.
+     * @return A new instance of fragment fragment_player_soccer.
      */
     // TODO: Rename and change types and number of parameters
-    public static fragment_team_soccer newInstance(String param1, String param2) {
-        fragment_team_soccer fragment = new fragment_team_soccer();
+    public static fragment_player_soccer newInstance(String param1, String param2) {
+        fragment_player_soccer fragment = new fragment_player_soccer();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +60,6 @@ public class fragment_team_soccer extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_team_soccer, container, false);
+        return inflater.inflate(R.layout.fragment_player_soccer, container, false);
     }
 }
