+++ swProject/app/src/main/java/com/example/swproject/ui/SoccerRankingActivity.java
@@ -1,8 +1,11 @@
 package com.example.swproject.ui;
 
 import android.annotation.SuppressLint;
+import android.content.Intent;
 import android.os.AsyncTask;
 import android.os.Bundle;
+import android.view.View;
+import android.widget.ImageButton;
 import android.widget.TextView;
 
 import androidx.annotation.Nullable;
@@ -48,6 +51,16 @@ public class SoccerRankingActivity extends soccerActivity {
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.soccer_ranking);
+
+        /* home 아이콘 눌렀을 때 메인화면 */
+        ImageButton imageButton = (ImageButton) findViewById(R.id.homeicon);
+        imageButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(getApplicationContext(),MainActivity.class);
+                startActivity(intent);
+            }
+        });
 
         team_data_ = findViewById(R.id.text_team_ranking);
         team_data_1 = findViewById(R.id.text_team_ranking1);
