+++ swProject/app/src/main/java/com/example/swproject/ui/SoccerScheduleActivity.java
@@ -1,9 +1,12 @@
 package com.example.swproject.ui;
 
 import android.annotation.SuppressLint;
+import android.content.Intent;
 import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
+import android.view.View;
+import android.widget.ImageButton;
 import android.widget.TextView;
 
 import androidx.annotation.Nullable;
@@ -33,6 +36,17 @@ public class SoccerScheduleActivity extends soccerActivity {
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.soccer_schedule);
+
+        /* home 아이콘 눌렀을 때 메인화면 */
+        ImageButton imageButton = (ImageButton) findViewById(R.id.homeicon);
+        imageButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(getApplicationContext(),MainActivity.class);
+                startActivity(intent);
+            }
+        });
+        /* home 아이콘 눌렀을 때 메인화면 */
 
         schedule_ = findViewById(R.id.text_schedule);
         result_ = findViewById(R.id.text_result);
