+++ swProject/app/src/main/java/com/example/swproject/ui/soccerActivity.java
@@ -4,6 +4,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+import android.widget.ImageButton;
 
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
@@ -15,6 +16,7 @@ public class soccerActivity extends AppCompatActivity {
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.soccer);
+
 
         Button btnNs = (Button) findViewById(R.id.btnNewsS);
         btnNs.setOnClickListener(new View.OnClickListener(){
