+++ swProject/app/src/main/res/layout/activity_main.xml
@@ -15,11 +15,7 @@
         android:orientation="horizontal"
         android:background="#1C6A1F"
         android:gravity="center_vertical">
-        <ImageView
-            android:layout_width="50dp"
-            android:layout_height="match_parent"
-            android:scaleType="center"
-            android:src="@drawable/ic_baseline_add_alert_24"/>
+
         <TextView
             android:layout_width="0dp"
             android:layout_height="wrap_content"
@@ -27,20 +23,16 @@
             android:text="sport alarm services"
             android:textSize="20dp"
             android:gravity="center"
+            android:layout_marginStart="40dp"
             android:textColor="@android:color/white"/>
-        <ImageView
-            android:id="@+id/alarm"
-            android:layout_width="50dp"
-            android:layout_height="match_parent"
-            android:scaleType="center"
-            android:src="@drawable/ic_baseline_more_horiz_24"/>
+
 
         <ImageView
             android:id="@+id/more"
             android:layout_width="50dp"
             android:layout_height="match_parent"
             android:scaleType="center"
-            android:src="@drawable/ic_baseline_home_24"/>
+            android:src="@drawable/ic_baseline_add_alert_24"/>
 
     </LinearLayout>
